+++ internal/sidecar/credstore_bench_test.go
@@ -25,6 +25,6 @@ func BenchmarkCredStoreSelectMixed(b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		_ = cs.Select(ProviderAnthropic)
+		_ = cs.Select(ProviderAnthropic, "")
 	}
 }